refactor: move pre-fetch to Manage Boxes, only preload for current base and most used states
